* finalized patching, fixes
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@284 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ protocolgame.cpp
@@ -1971,7 +1971,7 @@ void ProtocolGame::sendCloseShop()
 	}
 }
 
-void ProtocolGame::sendGoods(const std::map<uint16_t, uint8_t>& itemMap)
+void ProtocolGame::sendGoods(const std::map<uint32_t, uint32_t>& itemMap)
 {
 	NetworkMessage* msg = getOutputBuffer();
 	if(msg)
@@ -1982,7 +1982,7 @@ void ProtocolGame::sendGoods(const std::map<uint16_t, uint8_t>& itemMap)
 		msg->AddByte(std::min((size_t)255, itemMap.size()));
 
 		uint32_t i = 0;
-		for(std::map<uint16_t, uint8_t>::const_iterator it = itemMap.begin(); it != itemMap.end() && i < 255; ++it, ++i)
+		for(std::map<uint32_t, uint32_t>::const_iterator it = itemMap.begin(); it != itemMap.end() && i < 255; ++it, ++i)
 		{
 			msg->AddItemId(it->first);
 			msg->AddByte(it->second);
